20812: Migrate detail operations to abstract base class (#228)
- Moves detail/metric methods to base class and removes them from client
implementation
- Some typing fixes
