+++ howso/client/schemas/__init__.py
@@ -2,5 +2,5 @@ from .base import BaseSchema
 from .project import Project, ProjectDict
 from .reaction import Reaction
 from .session import Session, SessionDict
-from .trainee import Trainee, TraineeDict, TraineePersistence
+from .trainee import Trainee, TraineeDict
 from .version import HowsoVersion
