+++ howso/direct/core.py
@@ -640,45 +640,6 @@ class HowsoCore:
             "remove_duplicates": remove_duplicates,
         })
 
-    def set_substitute_feature_values(
-        self,
-        trainee_id: str,
-        substitution_value_map: Union[Dict, None]
-    ) -> None:
-        """
-        Set substitution feature values used in case generation.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        substitution_value_map : dict or None
-            A dictionary of feature name to value to substitution value. If the
-            map is None, all substitutions will be disabled and cleared.
-        """
-        return self._execute(
-            trainee_id, "set_substitute_feature_values",
-            {
-                "substitution_value_map": substitution_value_map,
-            }
-        )
-
-    def get_substitute_feature_values(self, trainee_id: str) -> Dict:
-        """
-        Get substitution feature values used in case generation.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-
-        Returns
-        -------
-        dict
-            The dictionary of feature name to value to substitution value.
-        """
-        return self._execute(trainee_id, "get_substitute_feature_values", {})
-
     def set_session_metadata(
         self,
         trainee_id: str,
@@ -762,86 +723,6 @@ class HowsoCore:
                 "session": session,
             }
         )
-
-    def remove_feature(
-        self,
-        trainee_id: str,
-        feature: str,
-        *,
-        condition: Optional[Dict] = None,
-        condition_session: Optional[str] = None,
-        session: Optional[str] = None
-    ) -> None:
-        """
-        Remove a feature.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        feature : str
-            The feature name.
-        condition : str, optional
-            A condition map where features will only be removed when certain
-            criteria is met.
-        condition_session : str optional
-            If specified, ignores the condition parameter and operates on cases
-            for the specified session id.
-        session : str, optional
-            The identifier of the Trainee session to associate the feature
-            removal with.
-        """
-        return self._execute(trainee_id, "remove_feature", {
-            "feature": feature,
-            "condition": condition,
-            "session": session,
-            "condition_session": condition_session,
-        })
-
-    def add_feature(
-        self,
-        trainee_id: str,
-        feature: str,
-        feature_value: Optional[Union[int, float, str]] = None,
-        *,
-        condition: Optional[Dict] = None,
-        condition_session: Optional[str] = None,
-        feature_attributes: Optional[Dict] = None,
-        overwrite: bool = False,
-        session: Optional[str] = None,
-    ) -> None:
-        """
-        Add a feature.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        feature : str
-            The feature name.
-        feature_value : int or float or str, optional
-            The feature value.
-        condition : str, optional
-            A condition map where features will only be removed when certain
-            criteria is met.
-        condition_session : str optional
-            If specified, ignores the condition parameter and operates on cases
-            for the specified session id.
-        overwrite : bool, default False
-            If True, the feature will be over-written if it exists.
-        session : str, optional
-            The identifier of the Trainee session to associate the feature
-            addition with.
-        """
-        return self._execute(trainee_id, "add_feature", {
-            "feature": feature,
-            "feature_value": feature_value,
-            "overwrite": overwrite,
-            "condition": condition,
-            "feature_attributes": feature_attributes,
-            "session": session,
-            "condition_session": condition_session,
-        })
 
     def get_auto_ablation_params(self, trainee_id: str):
         """Get trainee parameters for auto-ablation set by :meth:`set_auto_ablation_params`."""
@@ -2191,62 +2072,6 @@ class HowsoCore:
             "weight_feature": weight_feature,
         })
 
-    def compute_conviction_of_features(
-        self,
-        trainee_id: str,
-        *,
-        features: Optional[Iterable[str]] = None,
-        action_features: Optional[Iterable[str]] = None,
-        familiarity_conviction_addition: bool = True,
-        familiarity_conviction_removal: bool = False,
-        weight_feature: Optional[str] = None,
-        use_case_weights: bool = False
-    ) -> Dict:
-        """
-        Get familiarity conviction for features in the model.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The id of the trainee.
-        features : iterable of str, optional
-            An iterable of feature names to calculate convictions. At least 2
-            features are required to get familiarity conviction. If not
-            specified all features will be used.
-        action_features : iterable of str, optional
-            An iterable of feature names to be treated as action features
-            during conviction calculation in order to determine the conviction
-            of each feature against the set of action_features. If not
-            specified, conviction is computed for each feature against the rest
-            of the features as a whole.
-        familiarity_conviction_addition : bool, default True
-            Calculate and output familiarity conviction of adding the
-            specified features in the output.
-        familiarity_conviction_removal : bool, default False
-            Calculate and output familiarity conviction of removing
-            the specified features in the output.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        use_case_weights : bool, default False
-            If set to True will scale influence weights by each
-            case's weight_feature weight.
-
-        Returns
-        -------
-        dict
-            A dict with familiarity_conviction_addition or
-            familiarity_conviction_removal.
-        """
-        return self._execute(trainee_id, "compute_conviction_of_features", {
-            "features": features,
-            "action_features": action_features,
-            "familiarity_conviction_addition": familiarity_conviction_addition,
-            "familiarity_conviction_removal": familiarity_conviction_removal,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-        })
-
     def get_session_indices(self, trainee_id: str, session: str) -> List[int]:
         """
         Get list of all session indices for a specified session.
@@ -2286,61 +2111,6 @@ class HowsoCore:
         return self._execute(trainee_id, "get_session_training_indices", {
             "session": session,
         })
-
-    def set_internal_parameters(self, trainee_id: str, params: Dict) -> None:
-        """
-        Sets specific model parameters in the Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        params : dict
-            A dictionary containing the internal parameters.
-        """
-        return self._execute(trainee_id, "set_internal_parameters", {
-            **params
-        })
-
-    def set_feature_attributes(
-        self,
-        trainee_id: str,
-        feature_attributes: Dict[str, Dict]
-    ) -> Dict[str, Dict]:
-        """
-        Sets feature attributes for a Trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        feature_attributes : dict of str to dict
-            A dictionary of feature name to dictionary of feature attributes.
-
-        Returns
-        -------
-        dict
-            The updated feature attributes.
-        """
-        return self._execute(trainee_id, "set_feature_attributes", {
-            "features": feature_attributes,
-        })
-
-    def get_feature_attributes(self, trainee_id: str) -> Dict[str, Dict]:
-        """
-        Get Trainee feature attributes.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee
-
-        Returns
-        -------
-        dict
-            A dictionary of feature name to dictionary of feature attributes.
-        """
-        return self._execute(trainee_id, "get_feature_attributes", {})
 
     def export_trainee(
         self,
@@ -2410,275 +2180,6 @@ class HowsoCore:
         """
         params = {**kwargs}
         return self._execute(trainee_id, "analyze", params)
-
-    def get_marginal_stats(
-        self,
-        trainee_id: str,
-        *,
-        condition: Optional[Dict[str, Any]] = None,
-        num_cases: Optional[int] = None,
-        precision: Optional[Literal["exact", "similar"]] = None,
-        weight_feature: Optional[str] = None
-    ) -> Dict[str, Dict[str, float]]:
-        """
-        Get marginal stats for all features.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        condition : dict or None, optional
-            A condition map to select which cases to compute marginal stats
-            for.
-
-            .. NOTE::
-                The dictionary keys are the feature name and values are one of:
-
-                    - None
-                    - A value, must match exactly.
-                    - An array of two numeric values, specifying an inclusive
-                      range. Only applicable to continuous and numeric ordinal
-                      features.
-                    - An array of string values, must match any of these values
-                      exactly. Only applicable to nominal and string ordinal
-                      features.
-        num_cases : int, default None
-            The maximum amount of cases to use to calculate marginal stats.
-            If not specified, the limit will be k cases if precision is
-            "similar". Only used if `condition` is not None.
-        precision : str, default None
-            The precision to use when selecting cases with the condition.
-            Options are 'exact' or 'similar'. If not specified "exact" will be
-            used. Only used if `condition` is not None.
-        weight_feature : str, optional
-            When specified, will attempt to return stats that were computed
-            using this weight_feature.
-
-        Returns
-        -------
-        dict of str to dict of str to float
-            A map of feature names to map of stat type to stat values.
-        """
-        return self._execute(trainee_id, "get_marginal_stats", {
-            "condition": condition,
-            "num_cases": num_cases,
-            "precision": precision,
-            "weight_feature": weight_feature,
-        })
-
-    def get_internal_parameters(
-        self,
-        trainee_id: str,
-        *,
-        action_feature: Optional[str] = None,
-        context_features: Optional[Iterable[str]] = None,
-        mode: Optional[Literal["robust", "full"]] = None,
-        weight_feature: Optional[str] = None,
-    ) -> Dict:
-        """
-        Get the parameters used by the Trainee.
-
-        If 'action_feature', 'context_features', 'mode', or 'weight_feature'
-        are specified, then the best hyperparameters analyzed in the Trainee
-        are the value of the 'hyperparameter_map' key, otherwise this value
-        will be the dictionary containing all the hyperparameter sets in
-        the Trainee.
-
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        action_feature : str, optional
-            If specified will return the best analyzed hyperparameters to
-            target this feature.
-        context_features : str, optional
-            If specified, will find and return the best analyzed hyperparameters
-            to use with these context features.
-        mode : str, optional
-            If specified, will find and return the best analyzed hyperparameters
-            that were computed in this mode.
-        weight_feature : str, optional
-            If specified, will find and return the best analyzed hyperparameters
-            that were analyzed using this weight feaure.
-
-        Returns
-        -------
-        dict
-            A dict including the either all of the Trainee's internal
-            parameters or only the best hyperparameters selected using the
-            passed parameters.
-        """
-        return self._execute(trainee_id, "get_internal_parameters", {
-            "action_feature": action_feature,
-            "context_features": context_features,
-            "mode": mode,
-            "weight_feature": weight_feature,
-        })
-
-    def pairwise_distances(
-        self,
-        trainee_id: str,
-        features: Optional[Iterable[str]] = None,
-        *,
-        action_feature: Optional[str] = None,
-        from_case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        from_values: Optional[List[List[Any]]] = None,
-        to_case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        to_values: Optional[List[List[Any]]] = None,
-        use_case_weights: bool = False,
-        weight_feature: Optional[str] = None
-    ) -> List:
-        """
-        Compute pairwise distances between specified cases.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        features : iterable of str, optional
-            List of feature names to use when computing pairwise distances.
-            If unspecified uses all features.
-        action_feature : str, optional
-            The action feature. If specified, uses targeted hyperparameters
-            used to predict this `action_feature`, otherwise uses targetless
-            hyperparameters.
-        from_case_indices : Iterable of Sequence[Union[str, int]], optional
-            An iterable of sequences, of session id and index, where index
-            is the original 0-based index of the case as it was trained into
-            the session. If specified must be either length of 1 or match
-            length of `to_values` or `to_case_indices`.
-        from_values : list of list of object, optional
-            A 2d-list of case values. If specified must be either length of
-            1 or match length of `to_values` or `to_case_indices`.
-        to_case_indices : Iterable of Sequence[Union[str, int]], optional
-            An Iterable of Sequences, of session id and index, where index
-            is the original 0-based index of the case as it was trained into
-            the session. If specified must be either length of 1 or match
-            length of `from_values` or `from_case_indices`.
-        to_values : list of list of object, optional
-            A 2d-list of case values. If specified must be either length of
-            1 or match length of `from_values` or `from_case_indices`.
-        use_case_weights : bool, default False
-            If set to True, will scale influence weights by each case's
-            `weight_feature` weight.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-
-        Returns
-        -------
-        list
-            A list of computed pairwise distances between each corresponding
-            pair of cases in `from_case_indices` and `to_case_indices`.
-        """
-        return self._execute(trainee_id, "pairwise_distances", {
-            "features": features,
-            "action_feature": action_feature,
-            "from_case_indices": from_case_indices,
-            "from_values": from_values,
-            "to_case_indices": to_case_indices,
-            "to_values": to_values,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-        })
-
-    def distances(
-        self,
-        trainee_id: str,
-        features: Optional[Iterable[str]] = None,
-        *,
-        action_feature: Optional[str] = None,
-        case_indices: Optional[Iterable[Sequence[Union[str, int]]]] = None,
-        feature_values: Optional[Iterable[Any]] = None,
-        use_case_weights: bool = False,
-        weight_feature: Optional[str] = None,
-        row_offset: int = 0,
-        row_count: Optional[int] = None,
-        column_offset: int = 0,
-        column_count: Optional[int] = None
-    ) -> Dict:
-        """
-        Compute distances matrix for specified cases.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        features : iterable of str, optional
-            List of feature names to use when computing distances. If
-            unspecified uses all features.
-        action_feature : str, optional
-            The action feature. If specified, uses targeted hyperparameters
-            used to predict this `action_feature`, otherwise uses targetless
-            hyperparameters.
-        case_indices : Iterable of Sequence[Union[str, int]], optional
-            An Iterable of Sequences, of session id and index, where index is
-            the original 0-based index of the case as it was trained into the
-            session. If specified, returns distances for all of these
-            cases. Ignored if `feature_values` is provided. If neither
-            `feature_values` nor `case_indices` is specified, uses full dataset.
-        feature_values : list of object, optional
-            If specified, returns distances of the local model relative to
-            these values, ignores `case_indices` parameter.
-        use_case_weights : bool, default False
-            If set to True, will scale influence weights by each case's
-            `weight_feature` weight.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-
-        Returns
-        -------
-        dict
-            A dictionary of keys, 'distances', 'row_case_indices' and
-            'column_case_indices'.
-        """
-        return self._execute(trainee_id, "distances", {
-            "features": features,
-            "action_feature": action_feature,
-            "case_indices": case_indices,
-            "feature_values": feature_values,
-            "weight_feature": weight_feature,
-            "use_case_weights": use_case_weights,
-            "row_offset": row_offset,
-            "row_count": row_count,
-            "column_offset": column_offset,
-            "column_count": column_count,
-        })
-
-    def evaluate(
-        self,
-        trainee_id: str,
-        features_to_code_map: Dict[str, str],
-        *,
-        aggregation_code: Optional[str] = None
-    ) -> Dict:
-        """
-        Evaluate custom code on feature values of all cases in the trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        features_to_code_map : dict of str to str
-            A dictionary with feature name keys and custom Amalgam code string
-            values.
-        aggregation_code : str, optional
-            A string of custom Amalgam code that can access the list of values
-            derived form the custom code in features_to_code_map.
-
-        Returns
-        -------
-        dict
-            A dictionary with keys: 'evaluated' and 'aggregated'.
-        """
-        return self._execute(
-            trainee_id, "evaluate",
-            {
-                "features_to_code_map": features_to_code_map,
-                "aggregation_code": aggregation_code
-            })
 
     def reset_parameter_defaults(self, trainee_id: str) -> None:
         """
